reverted some of the trailing return types

+++ src/dbg/input-manager.hpp
@@ -12,5 +12,5 @@ public:
     ) const -> bool;
     auto getMousePosition(
         sf::RenderWindow &window
-    ) const -> sf::Vector2i;
+    ) const -> sf::Vector2f;
 };
